feat: Use fontsource instead of restribution of fonts and icons (#96)
## Description

Use fontsource to install the fonts via npm instead of manually
redistributing the fonts and icons. The benefit is that fonts can
receive updates and icons don't have to be uploaded manually.

To use icons, you have to find out the name of the icon
(https://fonts.google.com/icons). Then, there are two different ways:
- As partial in html templates: 
  ```go
  {{ partial "icon" "<icon-name>" }}
  ```
- As shortcode in markdown templates: 
  ```go
  {{< icon <icon-name> >}}
  ```

The only disadvantage is the size of 270kB, but it's lazy-loaded and we
have all icons.

It might be good to add the shortcut to the contribution wiki after the
PR has been merged.

For the Roboto font, the improvement is even more significant (esp. for
Chrome users). It recognizes the font as actual Roboto font and loads it
from chrome resources, reducing the size of ~300kB to ~200 bytes.

TODO: 

- [ ] Change outlined to rounded icons