feat: Quick-start for new pages using improved archetypes (#139)
The previous usage of archetypes wasn't properly documented. In
addition, the generated content raised syntax errors.

I rewrote the archetypes to use HTML comments to describe the type of
content that is expected. The comments are stripped in the production
environment with the `--minify` flag. The main advantage is that it's
easier to get started and the page can be shared earlier since it
doesn't contain any generic content anymore.

To try it out, follow the new instructions in the `CONTRIBUTING.md`
(should be part of the wiki too).

Co-authored-by: Lennart Rommeiss <61516567+lenderom@users.noreply.github.com>