refactor: Replace "ico" partial with "logo" partial (#140)
The icon usage was replaced by fontsource in #96. The remaining icon
infrastructure was only used for the logos of operators.

This PR simplifies the code by replacing the "ico" partial with a "logo"
partial.

In addition to the high complexity of the old "_ico.html", it had some
issues:
- Multiple return statements were used in one partial. This is not
allowed as per docs ("Only one return statement is allowed per partial
file."): https://gohugo.io/templates/partial/#example-getimage
- The whole build process broke when an operator was added without
related logo.

---------

Signed-off-by: lennartrommeiss <lennart.rommeiss@deutschebahn.com>
Co-authored-by: Lennart Rommeiss <61516567+lenderom@users.noreply.github.com>
Co-authored-by: lennartrommeiss <lennart.rommeiss@deutschebahn.com>