fix: Close search when clicking outside of search area (#210)
So far, the search dialog could be closed by clicking outside of the
search element, but only if no text was entered. If text was entered in
the input field, it was no longer possible. This commit closes the
search on click outside, independent of entered text in the search
field.

Co-authored-by: Lennart Rommeiss <61516567+lenderom@users.noreply.github.com>