refactor: Rename `.o-list--columns` to `.o-section--columns` (#213)
As correctly mentioned by @therobrob in
https://github.com/fipguide/fipguide.github.io/pull/180#discussion_r2156905739,
`.o-list` is already used for the overview pages and we should change
the name of `.o-list--columns`. Since it's actually not just usable by
lists itself and can be used to split all different kinds of sections
into columns, `.o-section--columns` is a good replacement in my eyes.

Co-authored-by: Robert Schuster <77234379+therobrob@users.noreply.github.com>