refactor: Replace boostrap with custom CSS (#208)
Bootstrap was a good library to get started faster, but we've mostly
used custom styling and most of the bootstrap functionality wasn't used.

Removing bootstrap provides a few advantages:
- We gain full control over the styles
- Better adoption to CSS BEM (we can choose our own class names)
- Reduced bundle size
- Better DevTools support. Chrome and Firefox have some neat development
tools around grid layout. The devtools didn't work with the bootstrap
grid layout.
- Less chance of breaking changes. Bootstrap is under active development
and can introduce breaking changes. CSS itself is considered as rather
stable.

As proposed in
https://github.com/fipguide/fipguide.github.io/pull/192#discussion_r2156820858,
the removal of bootstrap as separate PR. Resolves #74.

Beside some minor changes to margins and sizes, the most significant
visual change is the style of tables, which I changed as part of this
PR.

---------

Co-authored-by: Robert Schuster <robert.r.schuster@deutschebahn.com>