docs: Remove ignore cache instructions (#220)
The cache is mainly used by images and is generally not bad. With an
increasing number of images, the preprocessing takes longer, leading to
an increased startup time.

There are some caching issues with hugo, but these are not solved using
the `--ignoreCache` attribute. Instead, a full restart of the
application is necessary in certain cases.