fix: Avoid layout shifts by pre-allocating space for pictures
So far, images don't set a fixed height for images. This leads to
layout shifts while loading since CSS can only recognize the size
of the image when loaded.

By setting the aspact ratio on the img-element, the browser can
already pre-allocate the necessary space. When the image is loaded,
the layout no longer shifts.