feat: Add interactive map to navigate countries (#251)
Interactive map that can be used
to navigate the countries on the countries page, in addition to the list
of countries.

It's the foundation for the map and can be extended by indicating the
FIP rating via background color, show additional information on hover or
click etc.

Resolves #135

In a later step, we could optimize the behavior and add the CSS classes
to available countries in the SVG during build time, not with JavaScript
during run time. However, this is more complex, we need to parse the SVG
in Go templates.