feat: Add instructions for AI agents (#393)
Add an [`AGENTS.md`](https://agents.md/) file for instructions for AI
agents. The file is understood by Github Copilot, Claude Code, and other
agents.

In addition, add "translate" and "check" commands for pages using the
Opencode framework. OpenCode is open source and can be used with
different providers and models (OpenAI, Github Copilot, Anthropic).
Instructions how to use the custom commands are available here:
https://github.com/fipguide/fipguide.github.io/blob/feat/opencode/CONTRIBUTING.md#use-ai-for-translation-and-quality-checks

It's not perfect yet - but we can improve the prompts and glossary over
time.

Resolves #161

---------

Signed-off-by: lennartrommeiss <61516567+lenderom@users.noreply.github.com>
Co-authored-by: lennartrommeiss <61516567+lenderom@users.noreply.github.com>