fix: Allow closing the search via the search button (#424)
Especially on mobile devices, space is limited and users may not have
enough space to close the search if they want to. This PR changes the
"open search" behavior to a "toggle search" behavior (open if search is
closed, close if search is open).

In addition, two bugs are fixed:
- The overlay wasn't displayed when the user navigated to the search
field on the home page via tab.
- The overlay wasn't closed if the search was opened via click and then
moved out using the tab character (this will probably change when we
implement the focus trapping, but for now it's the best solution I
think).

---------

Co-authored-by: Robert Schuster <77234379+therobrob@users.noreply.github.com>