feat: Add train category expander (#405)
This PR introduces a new train category expander. The goal is a better
consistency across pages and a better explanation of the icons used in
the expander summary.

Part of the PR is also a refactoring of the text highlight styles to use
the same logic for tags and text highlights.

To avoid too many merge conflicts, we should merge the technical
implementation before adjusting all country pages as this might take
some time. A helper has been added to the opencode commands and can by
used by invoking `/update-traincategory {operator_name}`.